--- lib/helpers/RouteNames.dart
@@ -1,9 +1,0 @@
-class RouteNames{
-  static final rotaAddTarefaNoFuncionarioPage = '/AddTarefaNoFuncionarioPage';
-  static final rotaStartPage = '/StartPage';
-  static final rotaAddTarefaPage = '/AddTarefaPage';
-  static final rotaRegisterPage = '/RegisterPage';
-  static final rotaListFuncionarioPage = '/ListFuncionarioPage';
-  static final rotaListTarefaPage = '/ListTarefaPage';
-  static final rotaAddFuncionariosPage = '/AddFuncionariosPage';
- }
