+++ lib/helpers/route_names.dart
@@ -1,0 +1,9 @@
+class RouteNames{
+  static const rotaAddTarefaNoFuncionarioPage = '/AddTarefaNoFuncionarioPage';
+  static const rotaStartPage = '/StartPage';
+  static const rotaAddTarefaPage = '/AddTarefaPage';
+  static const rotaRegisterPage = '/RegisterPage';
+  static const rotaListFuncionarioPage = '/ListFuncionarioPage';
+  static const rotaListTarefaPage = '/ListTarefaPage';
+  static const rotaAddFuncionariosPage = '/AddFuncionariosPage';
+ }
