+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:banco/Pages/AppRouter.dart';
+import 'package:banco/Pages/app_router.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 import 'helpers/database_helper.dart';
 
@@ -9,5 +9,5 @@ Future<void> main() async {
   runApp( const AppRouter());
   var db = DatabaseHelper();
   await db.initConection();
-  await db.CreateTable();
+  await db.createTable();
 }
