+++ lib/models/Funcionario.dart
@@ -9,15 +9,15 @@ class Funcionario{
 
   Future<int?>insert() async{
     Database db = await database!;
-    var resultado = await db.insert(_funcinaroTableName,this.toMap());
+    var resultado = await db.insert(_funcinaroTableName,toMap());
     return resultado;
   }
 
   Future<int>update() async{
     var db = await database!;
-    var resultado = await db.update(_funcinaroTableName,this.toMap(),
+    var resultado = await db.update(_funcinaroTableName,toMap(),
       where: "Id = ?",
-      whereArgs: [this.id]);
+      whereArgs: [id]);
     return resultado;
   }
 
@@ -25,7 +25,7 @@ class Funcionario{
     var db = await database!;
     var resultado = await db.delete(_funcinaroTableName,
       where: "Id = ?",
-      whereArgs: [this.id]);
+      whereArgs: [id]);
     return resultado;
   }
 
@@ -44,23 +44,21 @@ class Funcionario{
     _apto = map['Apto'];
   }
 
-  late String CreateScript = '''CREATE TABLE IF NOT EXISTS Funcionario (
+  late String createScript = '''CREATE TABLE IF NOT EXISTS Funcionario (
                                   Id INTEGER PRIMARY KEY AUTOINCREMENT,
                                   Nome TEXT,
                                   Apto INTEGER
                                 )''';
                                 
-  late String DropScript = 'DROP TABLE IF EXISTS Funcionario';
+  late String dropScript = 'DROP TABLE IF EXISTS Funcionario';
    //Fim
 
   late int? id;
   late String nome;
   late int _apto;
 
-  Funcionario({int? id = null,String nome = '' , bool apto = true}){
-    this.id = id;
-    this.nome = nome;
-    this._apto = (apto == true) ? 1 : 0;
+  Funcionario({this.id,this.nome = '' , bool apto = true}){
+    _apto = (apto == true) ? 1 : 0;
   }
 
   bool get apto{
