+++ lib/models/stop_watch_tarefa.dart
@@ -8,7 +8,7 @@ class StopWatchTarefa{
 
   Future<int?>insert() async{
     Database db = await database!;
-    var resultado = await db.insert(_tarefaTableNome,this.toMap());
+    var resultado = await db.insert(_tarefaTableNome,toMap());
     return resultado;
   }
 
@@ -17,8 +17,8 @@ class StopWatchTarefa{
     Database db = await database!;
     var map = <String,dynamic>{
       'start_timestamp' : 'datetime("now")',
-      'id_funcionario' : id_funcionario,
-      'id_tarefa' :id_tarefa,
+      'id_funcionario' : idFuncionario,
+      'id_tarefa' :idTarefa,
     };
     var resultado = await db.insert(_tarefaTableNome,map);
     return resultado;
@@ -28,29 +28,29 @@ class StopWatchTarefa{
     Database db = await database!;
     var resultado = await db.update(_tarefaTableNome,{'stop_timestamp': 'datetime("now")'},
       where: "id_funcionario = ? and stop_timestamp is null",
-      whereArgs: [this.id_funcionario]);
+      whereArgs: [idFuncionario]);
     return resultado;
   }
 
   Map<String,dynamic> toMap(){
     var map = <String,dynamic>{
-      'start_timestamp' : start_timestamp,
-      'stop_timestamp'  : stop_timestamp,
-      'id_funcionario' : id_funcionario,
-      'id_tarefa' :id_tarefa,
+      'start_timestamp' : startTimestamp,
+      'stop_timestamp'  : stopTimestamp,
+      'id_funcionario' : idFuncionario,
+      'id_tarefa' :idTarefa,
     };
     return map;
   }
 
   StopWatchTarefa.fromMap(Map<String,dynamic> map){
-    start_timestamp = map['start_timestamp'];
-    stop_timestamp = map['stop_timestamp'];
-    id_funcionario = map['id_funcionario'];
-    id_tarefa = map['id_tarefa'];
+    startTimestamp = map['start_timestamp'];
+    stopTimestamp = map['stop_timestamp'];
+    idFuncionario = map['id_funcionario'];
+    idTarefa = map['id_tarefa'];
   }
 
 
-  late String CreateScript = '''CREATE TABLE IF NOT EXISTS StopWatchTarefa (
+  late String createScript = '''CREATE TABLE IF NOT EXISTS StopWatchTarefa (
                                   Id INTEGER PRIMARY KEY AUTOINCREMENT,
                                   start_timestamp DATETIME,
                                   stop_timestamp DATETIME,
@@ -58,20 +58,15 @@ class StopWatchTarefa{
                                   id_tarefa INTEGER
                                 )''';
                                 
-  late String DropScript = 'DROP TABLE IF EXISTS StopWatchTarefa';
+  late String dropScript = 'DROP TABLE IF EXISTS StopWatchTarefa';
 
    //Fim
 
   late int id;
-  late DateTime? start_timestamp;
-  late DateTime? stop_timestamp;
-  late int? id_funcionario;
-  late int? id_tarefa;
+  late DateTime? startTimestamp;
+  late DateTime? stopTimestamp;
+  late int? idFuncionario;
+  late int? idTarefa;
 
-  StopWatchTarefa({DateTime? start_timestamp , DateTime? stop_timestamp , int? id_funcionario = null , int? id_tarefa = null}){
-    this.start_timestamp = start_timestamp;
-    this.stop_timestamp = stop_timestamp;
-    this.id_funcionario = id_funcionario;
-    this.id_tarefa = id_tarefa;
-  }
+  StopWatchTarefa({this.startTimestamp , this.stopTimestamp , this.idFuncionario , this.idTarefa});
 }
